chore : move latihan soal component in modul to upper components and rename QuizController to CBT
