+++ src/app/student/_components/CBT/InformationQuiz.tsx
